Alpha v0.7
Made baseXP a constant and lowered it to 2.
Added xpMultiplier to PlayerProfile Model.
Multiplier is used to boost player xp in GameScene.
On level complete, set level as highest level complete with
sceneManager.
Added a blank case for level 0 (endless) spawns.
Registered profile saving when app is put in the background.
Pass player profile to MenuScene.
Added reading and writing multiplier to defaults.
Added a method to reset the player’s profile, keep the endless hi score
and increment the xpMultiplier.
Added variables in level select to determine which levels to draw and
whether to draw endless mode.
Added reset label in settings (locked until level 25)
Implemented profile reset, which then brings the player to the main
menu.